Share the admin superuser_client fixture, let TraceFactory set a timestamp
Trace.timestamp is auto_now_add, so tests can't place a trace in a window
without it.
